requirements.txt
google-generativeai
pymupdf
opencv-python
numpy
pillow
